Add direct IO network interface
The aim is to add a network interface which can be serviced just using
two callbacks. Something like that would allow libmav to be integrated
into MAVSDK.

The current implementation seems to hang somewhere and never actually
move on. I'm not sure yet why. Any hints would be most welcome.

Signed-off-by: Julian Oes <julian@oes.ch>